refactor: 피드백 반영 - 세션 스토어 분리
